+++ docs/changelog.rst
@@ -143,6 +143,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - When remembering OS window sizes for full screen windows use the size before
   the window became fullscreen (:iss:`4221`)
 
+- macOS: Fix keyboard input not working after toggling traditional fullscreen
+  till the window is clicked in
+
 
 0.23.1 [2021-08-17]
 ----------------------